[#32339] fix multi-command files test with --ctest option
--HG--
branch : v15

+++ run_aster/run_aster_main.py
@@ -294,7 +294,7 @@ def main(argv=None):
         args.test = True
         basename = osp.splitext(osp.basename(args.export))[0]
         output = osp.abspath(basename + ".mess")
-        if osp.isfile(output):
+        if osp.isfile(output) and need_split:
             os.remove(output)
         add = {15: "code"}
         for unit, typ in add.items():
